Separate extauth tests into their own file.
