+++ ambassador/tests/kat/t_extauth.py
@@ -1,0 +1,343 @@
+from kat.harness import Query
+
+from abstract_tests import AmbassadorTest, ServiceType, HTTP
+
+
+class AuthenticationHTTPBufferedTest(AmbassadorTest):
+
+    target: ServiceType
+    auth: ServiceType
+
+    def init(self):
+        self.target = HTTP()
+        self.auth = HTTP(name="auth")
+
+    def config(self):
+        yield self, self.format("""
+---
+apiVersion: ambassador/v0
+kind:  Module
+name:  ambassador
+config:
+  buffer:  
+    max_request_bytes: 16384
+    max_request_time: 5000
+---
+apiVersion: ambassador/v1
+kind: AuthService
+name:  {self.auth.path.k8s}
+proto: http
+auth_service: "{self.auth.path.k8s}"
+path_prefix: "/extauth"
+timeout_ms: 5000
+
+allowed_request_headers:
+- X-Foo
+- X-Bar
+- Requested-Status
+- Requested-Header
+- Location
+
+allowed_authorization_headers:
+- X-Foo
+
+allow_request_body: True
+""")
+        yield self, self.format("""
+---
+apiVersion: ambassador/v0
+kind:  Mapping
+name:  {self.target.path.k8s}
+prefix: /target/
+service: {self.target.path.k8s}
+""")
+
+    def queries(self):
+        # [0]
+        yield Query(self.url("target/"), headers={"Requested-Status": "401",
+                                                  "Baz": "baz",
+                                                  "Request-Header": "Baz"}, expected=401)
+        # [1]
+        yield Query(self.url("target/"), headers={"requested-status": "302",
+                                                  "location": "foo",
+                                                  "requested-header": "location"}, expected=302)
+        # [2]
+        yield Query(self.url("target/"), headers={"Requested-Status": "401",
+                                                  "X-Foo": "foo",
+                                                  "Requested-Header": "X-Foo"}, expected=401)
+        # [3]
+        yield Query(self.url("target/"), headers={"Requested-Status": "401",
+                                                  "X-Bar": "bar",
+                                                  "Requested-Header": "X-Bar"}, expected=401)
+        # [4]
+        yield Query(self.url("target/"), headers={"Requested-Status": "200",
+                                                  "Authorization": "foo-11111",
+                                                  "Requested-Header": "Authorization"}, expected=200)
+
+    def check(self):
+        # [0] Verifies all request headers sent to the authorization server.
+        assert self.results[0].backend.name == self.auth.path.k8s
+        assert self.results[0].backend.request.url.path == "/extauth/target/"
+        assert self.results[0].backend.request.headers["x-forwarded-proto"]== ["http"]
+        assert self.results[0].backend.request.headers["content-length"]== ["0"]
+        assert "x-forwarded-for" in self.results[0].backend.request.headers
+        assert "user-agent" in self.results[0].backend.request.headers
+        assert "baz" not in self.results[0].backend.request.headers
+        assert self.results[0].status == 401
+        assert self.results[0].headers["Server"] == ["envoy"]
+
+        # [1] Verifies that Location header is returned from Envoy.
+        assert self.results[1].backend.name == self.auth.path.k8s
+        assert self.results[1].backend.request.headers["requested-status"] == ["302"]
+        assert self.results[1].backend.request.headers["requested-header"] == ["location"]
+        assert self.results[1].backend.request.headers["location"] == ["foo"]
+        assert self.results[1].status == 302
+        assert self.results[1].headers["Server"] == ["envoy"]
+        assert self.results[1].headers["Location"] == ["foo"]
+
+        # [2] Verifies Envoy returns whitelisted headers input by the user.
+        assert self.results[2].backend.name == self.auth.path.k8s
+        assert self.results[2].backend.request.headers["requested-status"] == ["401"]
+        assert self.results[2].backend.request.headers["requested-header"] == ["X-Foo"]
+        assert self.results[2].backend.request.headers["x-foo"] == ["foo"]
+        assert self.results[2].status == 401
+        assert self.results[2].headers["Server"] == ["envoy"]
+        assert self.results[2].headers["X-Foo"] == ["foo"]
+
+        # [3] Verifies that envoy does not return not whitelisted headers.
+        assert self.results[3].backend.name == self.auth.path.k8s
+        assert self.results[3].backend.request.headers["requested-status"] == ["401"]
+        assert self.results[3].backend.request.headers["requested-header"] == ["X-Bar"]
+        assert self.results[3].backend.request.headers["x-bar"] == ["bar"]
+        assert self.results[3].status == 401
+        assert self.results[3].headers["Server"] == ["envoy"]
+        assert "X-Bar" not in self.results[3].headers
+
+        # [4] Verifies default whitelisted Authorization request header.
+        assert self.results[4].backend.request.headers["requested-status"] == ["200"]
+        assert self.results[4].backend.request.headers["requested-header"] == ["Authorization"]
+        assert self.results[4].backend.request.headers["authorization"] == ["foo-11111"]
+        assert self.results[4].status == 200
+        assert self.results[4].headers["Server"] == ["envoy"]
+        assert self.results[4].headers["Authorization"] == ["foo-11111"]
+
+
+class AuthenticationTestV1(AmbassadorTest):
+
+    target: ServiceType
+    auth: ServiceType
+
+    def init(self):
+        self.target = HTTP()
+        self.auth = HTTP(name="auth")
+
+    def config(self):
+        yield self, self.format("""
+---
+apiVersion: ambassador/v1
+kind: AuthService
+name:  {self.auth.path.k8s}
+auth_service: "{self.auth.path.k8s}"
+proto: http
+path_prefix: "/extauth"
+timeout_ms: 5000
+
+allowed_request_headers:
+- X-Foo
+- X-Bar
+- Requested-Status
+- Requested-Header
+- Location
+
+allowed_authorization_headers:
+- X-Foo
+
+
+""")
+        yield self, self.format("""
+---
+apiVersion: ambassador/v0
+kind:  Mapping
+name:  {self.target.path.k8s}
+prefix: /target/
+service: {self.target.path.k8s}
+""")
+
+    def queries(self):
+        # [0]
+        yield Query(self.url("target/"), headers={"Requested-Status": "401",
+                                                  "Baz": "baz",
+                                                  "Request-Header": "Baz"}, expected=401)
+        # [1]
+        yield Query(self.url("target/"), headers={"requested-status": "302",
+                                                  "location": "foo",
+                                                  "requested-header": "location"}, expected=302)
+        # [2]
+        yield Query(self.url("target/"), headers={"Requested-Status": "401",
+                                                  "X-Foo": "foo",
+                                                  "Requested-Header": "X-Foo"}, expected=401)
+        # [3]
+        yield Query(self.url("target/"), headers={"Requested-Status": "401",
+                                                  "X-Bar": "bar",
+                                                  "Requested-Header": "X-Bar"}, expected=401)
+        # [4]
+        yield Query(self.url("target/"), headers={"Requested-Status": "200",
+                                                  "Authorization": "foo-11111",
+                                                  "Requested-Header": "Authorization"}, expected=200)
+
+    def check(self):
+        # [0] Verifies all request headers sent to the authorization server.
+        assert self.results[0].backend.name == self.auth.path.k8s
+        assert self.results[0].backend.request.url.path == "/extauth/target/"
+        assert self.results[0].backend.request.headers["x-forwarded-proto"]== ["http"]
+        assert self.results[0].backend.request.headers["content-length"]== ["0"]
+        assert "x-forwarded-for" in self.results[0].backend.request.headers
+        assert "user-agent" in self.results[0].backend.request.headers
+        assert "baz" not in self.results[0].backend.request.headers
+        assert self.results[0].status == 401
+        assert self.results[0].headers["Server"] == ["envoy"]
+
+        # [1] Verifies that Location header is returned from Envoy.
+        assert self.results[1].backend.name == self.auth.path.k8s
+        assert self.results[1].backend.request.headers["requested-status"] == ["302"]
+        assert self.results[1].backend.request.headers["requested-header"] == ["location"]
+        assert self.results[1].backend.request.headers["location"] == ["foo"]
+        assert self.results[1].status == 302
+        assert self.results[1].headers["Server"] == ["envoy"]
+        assert self.results[1].headers["Location"] == ["foo"]
+
+        # [2] Verifies Envoy returns whitelisted headers input by the user.
+        assert self.results[2].backend.name == self.auth.path.k8s
+        assert self.results[2].backend.request.headers["requested-status"] == ["401"]
+        assert self.results[2].backend.request.headers["requested-header"] == ["X-Foo"]
+        assert self.results[2].backend.request.headers["x-foo"] == ["foo"]
+        assert self.results[2].status == 401
+        assert self.results[2].headers["Server"] == ["envoy"]
+        assert self.results[2].headers["X-Foo"] == ["foo"]
+
+        # [3] Verifies that envoy does not return not whitelisted headers.
+        assert self.results[3].backend.name == self.auth.path.k8s
+        assert self.results[3].backend.request.headers["requested-status"] == ["401"]
+        assert self.results[3].backend.request.headers["requested-header"] == ["X-Bar"]
+        assert self.results[3].backend.request.headers["x-bar"] == ["bar"]
+        assert self.results[3].status == 401
+        assert self.results[3].headers["Server"] == ["envoy"]
+        assert "X-Bar" not in self.results[3].headers
+
+        # [4] Verifies default whitelisted Authorization request header.
+        assert self.results[4].backend.request.headers["requested-status"] == ["200"]
+        assert self.results[4].backend.request.headers["requested-header"] == ["Authorization"]
+        assert self.results[4].backend.request.headers["authorization"] == ["foo-11111"]
+        assert self.results[4].status == 200
+        assert self.results[4].headers["Server"] == ["envoy"]
+        assert self.results[4].headers["Authorization"] == ["foo-11111"]
+
+        # TODO(gsagula): Write tests for all UCs which request header headers
+        # are overridden, e.g. Authorization.
+
+
+class AuthenticationTest(AmbassadorTest):
+    target: ServiceType
+    auth: ServiceType
+
+    def init(self):
+        self.target = HTTP()
+        self.auth = HTTP(name="auth")
+
+    def config(self):
+        yield self, self.format("""
+---
+apiVersion: ambassador/v0
+kind: AuthService
+name:  {self.auth.path.k8s}
+auth_service: "{self.auth.path.k8s}"
+path_prefix: "/extauth"
+
+allowed_headers:
+- X-Foo
+- X-Bar
+- Requested-Status
+- Requested-Header
+- Location
+- X-Foo
+
+""")
+        yield self, self.format("""
+---
+apiVersion: ambassador/v0
+kind:  Mapping
+name:  {self.target.path.k8s}
+prefix: /target/
+service: {self.target.path.k8s}
+""")
+
+    def queries(self):
+        # [0]
+        yield Query(self.url("target/"), headers={"Requested-Status": "401",
+                                                  "Baz": "baz",
+                                                  "Request-Header": "Baz"}, expected=401)
+        # [1]
+        yield Query(self.url("target/"), headers={"requested-status": "302",
+                                                  "location": "foo",
+                                                  "requested-header": "location"}, expected=302)
+        # [2]
+        yield Query(self.url("target/"), headers={"Requested-Status": "401",
+                                                  "X-Foo": "foo",
+                                                  "Requested-Header": "X-Foo"}, expected=401)
+        # [3]
+        yield Query(self.url("target/"), headers={"Requested-Status": "401",
+                                                  "X-Bar": "bar",
+                                                  "Requested-Header": "X-Bar"}, expected=401)
+        # [4]
+        yield Query(self.url("target/"), headers={"Requested-Status": "200",
+                                                  "Authorization": "foo-11111",
+                                                  "Requested-Header": "Authorization"}, expected=200)
+
+    def check(self):
+        # [0] Verifies all request headers sent to the authorization server.
+        assert self.results[0].backend.name == self.auth.path.k8s
+        assert self.results[0].backend.request.url.path == "/extauth/target/"
+        assert self.results[0].backend.request.headers["content-length"]== ["0"]
+        assert "x-forwarded-for" in self.results[0].backend.request.headers
+        assert "user-agent" in self.results[0].backend.request.headers
+        assert "baz" not in self.results[0].backend.request.headers
+        assert self.results[0].status == 401
+        assert self.results[0].headers["Server"] == ["envoy"]
+
+        # [1] Verifies that Location header is returned from Envoy.
+        assert self.results[1].backend.name == self.auth.path.k8s
+        assert self.results[1].backend.request.headers["requested-status"] == ["302"]
+        assert self.results[1].backend.request.headers["requested-header"] == ["location"]
+        assert self.results[1].backend.request.headers["location"] == ["foo"]
+        assert self.results[1].status == 302
+        assert self.results[1].headers["Server"] == ["envoy"]
+        assert self.results[1].headers["Location"] == ["foo"]
+
+        # [2] Verifies Envoy returns whitelisted headers input by the user.
+        assert self.results[2].backend.name == self.auth.path.k8s
+        assert self.results[2].backend.request.headers["requested-status"] == ["401"]
+        assert self.results[2].backend.request.headers["requested-header"] == ["X-Foo"]
+        assert self.results[2].backend.request.headers["x-foo"] == ["foo"]
+        assert self.results[2].status == 401
+        assert self.results[2].headers["Server"] == ["envoy"]
+        assert self.results[2].headers["X-Foo"] == ["foo"]
+
+        # [3] Verifies that envoy does not return not whitelisted headers.
+        assert self.results[3].backend.name == self.auth.path.k8s
+        assert self.results[3].backend.request.headers["requested-status"] == ["401"]
+        assert self.results[3].backend.request.headers["requested-header"] == ["X-Bar"]
+        assert self.results[3].backend.request.headers["x-bar"] == ["bar"]
+        assert self.results[3].status == 401
+        assert self.results[3].headers["Server"] == ["envoy"]
+        assert "X-Bar" in self.results[3].headers
+
+        # [4] Verifies default whitelisted Authorization request header.
+        assert self.results[4].backend.request.headers["requested-status"] == ["200"]
+        assert self.results[4].backend.request.headers["requested-header"] == ["Authorization"]
+        assert self.results[4].backend.request.headers["authorization"] == ["foo-11111"]
+        assert self.results[4].status == 200
+        assert self.results[4].headers["Server"] == ["envoy"]
+        assert self.results[4].headers["Authorization"] == ["foo-11111"]
+
+        # TODO(gsagula): Write tests for all UCs which request header headers
+        # are overridden, e.g. Authorization.
